// KLMPNHomeStay/ClientApp/src/app/model/secured/hs-member-details.ts
export class HsMemberDetails {
  hsId: string;
  hsName: string;
  //roomImage: string;

  hsImage1: string;
  hsImage2: string;
  hsImage3: string;
  hsImage4: string;
  hsImage5: string;
  hsImage6: string;
  hsImage7: string;
  hsImage8: string;
  hsImage9: string;
  hsImage10: string;

  roomImage1: string;
  roomImage2: string;
  roomImage3: string;
  roomImage4: string;
  roomImage5: string;
  roomImage6: string;
  roomImage7: string;
  roomImage8: string;
  roomImage9: string;
  roomImage10: string;

  latitude: string;
  longitude: string;
  roomDetailsList: any;
  roomImageStrings: string;
}

export class RoomDetails{
  roomCategory: string;
  roomFacility: any;
  roomNo: string;
  roomRate: string;
  roomFloor: string;
  noOfBeds: string;
  roomSize: string;
  isAvailable: string;
  roomImage: string;
}